Create Sum and count nodes
My target is to count all the nodes in a link list and find the sum of all the data present in nodes in a link list. For this I had created two functions  for counting-(One recursive and second using loops), I also created one function for sum using loops. If you are reading this, your target is to write recursive version of sum function

Happy Coding:)